frontend/requirements.txt
streamlit>=1.24.0
aiohttp>=3.8.0
pillow>=9.0.0
plotly>=5.13.0
pandas>=1.5.0
bcrypt>=4.0.0
altair>=4.2.0
numpy>=1.23.0
torch>=2.0.0
transformers>=4.30.0 